Prevent users fetching lists they don't own
On the client, we ensure that we don't attempt any other subscriptions,
unless the subsciption to fetch the requested lists succeeds. I.e. the
logged in user owns the list, etc. If the list subscription fails, we
simply display a proper error message and redirect the user to his/her
lists.