Fix bug when searching for duplicate list names
The bug causes a list named "Personal Stuff" to be considered the same
(i.e. a match) as another list named "Stuff". This is because the
regular expression I used failed to anchor the test to the start (^) and
end ($) of the string, hence causing "Stuff" to match "Personal Stuff"
even though Stuff was only occurring somewhere in the target string.

This update fixes this issue.