Fixed post-login buggy redirects (see full commit)
This update ensures that a user gets redirect to the page they were
looking at after a successful login. Before, users were getting
redirected to the top level route ('lists' for shoppers, and
'adminOverview' for admins) regardless of which page (or how many levels
deep in the route hierarchy) they were in before the page reload/logout
occurred.